move token stuff to token.rs, use bearer instead of basic
