+++ requests.http
@@ -5,7 +5,7 @@ GET http://127.0.0.1:8000/tasks/alive
 
 ###
 GET http://127.0.0.1:8000/task/1
-Authorization: Basic ed9e5e65-f899-4efa-82ad-04cfefb3d5ae
+Authorization: Bearer ed9e5e65-f899-4efa-82ad-04cfefb3d5ae
 
 ###
 
@@ -15,7 +15,7 @@ GET http://127.0.0.1:8000/task/20
 
 POST http://127.0.0.1:8000/task HTTP/1.1
 Content-Type: application/x-www-form-urlencoded
-Authorization: Basic ed9e5e65-f899-4efa-82ad-04cfefb3d5ae
+Authorization: Bearer ed9e5e65-f899-4efa-82ad-04cfefb3d5ae
 
 name=foo&project=frank&due=2021-07-25%2023%3A56%3A04
